SailBot simulator: first prototype working with bugs
Constant rotation of SailBot has a strange behaviour at 350-0 in JavaScript code, even though it seems to be working properly in Python. Heading gets stuck at 350 before going to 10 (when adding 10 each cycle).

Also, the sliders in the GUI have an overflow: when the slider is set to negative, the simulator receives 255+slider